Fixed fiber bug and some warnings
- isInitialized now work properly

+++ Scheduler/Source/MTDefaultAppInterop.cpp
@@ -127,8 +127,8 @@ namespace MT
 
 #elif MT_PLATFORM_POSIX || MT_PLATFORM_OSX
 
-		int pageSize = sysconf(_SC_PAGE_SIZE);
-		int pagesCount = size / pageSize;
+		int pageSize = (int)sysconf(_SC_PAGE_SIZE);
+		int pagesCount = (int)(size / pageSize);
 
 		//need additional page for stack tail
 		if ((size % pageSize) > 0)
